+++ docs/nuxt.config.ts
@@ -62,4 +62,6 @@ export default defineNuxtConfig({
       },
     },
   },
+
+  compatibilityDate: '2025-02-16',
 })