+++ packages/devtools/client/app.vue
@@ -9,9 +9,9 @@ import { useCopy } from '~/composables/editor'
 import { rpc } from '~/composables/rpc'
 import { registerCommands } from '~/composables/state-commands'
 import { splitScreenAvailable, splitScreenEnabled } from '~/composables/storage'
+import { useSchemaInput } from './composables/state-schema'
 import { useDevToolsOptions } from './composables/storage-options'
 import { setupClientRPC } from './setup/client-rpc'
-import { setupVueDevTools } from './setup/vue-devtools'
 
 import 'floating-vue/dist/style.css'
 import '@vue/devtools-applet/style.css'
@@ -38,8 +38,6 @@ useHead({
     },
   ],
 })
-
-setupVueDevTools()
 setupClientRPC()
 
 const client = useClient()
@@ -65,7 +63,7 @@ useEventListener('keydown', (e) => {
 })
 
 const { scale, sidebarExpanded } = useDevToolsOptions('ui')
-// const dataSchema = useSchemaInput()
+const dataSchema = useSchemaInput()
 
 onMounted(async () => {
   const injectClient = useInjectionClient()
@@ -157,8 +155,8 @@ registerCommands(() => [
     </div>
     <DisconnectIndicator />
     <RestartDialogs />
-    <!-- <div v-lazy-show="dataSchema">
+    <div v-lazy-show="dataSchema">
       <LazyDataSchemaDrawer />
-    </div> -->
+    </div>
   </div>
 </template>