+++ packages/devtools/client/composables/rpc.ts
@@ -1,7 +1,7 @@
 import type { ClientFunctions, ServerFunctions } from '../../src/types'
 import { useDebounce } from '@vueuse/core'
 import { createBirpc } from 'birpc'
-import { parse, stringify } from 'flatted'
+import { parse, stringify } from 'structured-clone-es'
 import { tryCreateHotContext } from 'vite-hot-client'
 import { ref, shallowRef } from 'vue'
 import { WS_EVENT_NAME } from '../../src/constant'
@@ -36,8 +36,15 @@ export const rpc = createBirpc<ServerFunctions, ClientFunctions>(clientFunctions
     const [namespace, fnName] = name.split(':') as [string, string]
     return extendedRpcMap.get(namespace)?.[fnName]
   },
-  onError(error, name) {
-    console.error(`[nuxt-devtools] RPC error on executing "${name}":`, error)
+  onFunctionError(error, name) {
+    console.error(`[nuxt-devtools] RPC error on executing "${name}":`)
+    console.error(error)
+    return true
+  },
+  onGeneralError(error) {
+    console.error(`[nuxt-devtools] RPC error:`)
+    console.error(error)
+    return true
   },
   timeout: 120_000,
 })
@@ -62,6 +69,7 @@ async function connectVite() {
 
   if (!hot) {
     wsConnecting.value = true
+    console.error('[nuxt-devtools] Unable to find Vite HMR context')
     throw new Error('Unable to connect to devtools')
   }
 
@@ -73,9 +81,13 @@ async function connectVite() {
   wsConnecting.value = true
 
   hot.on('vite:ws:connect', () => {
+    // eslint-disable-next-line no-console
+    console.log('[nuxt-devtools] Connected to WebSocket')
     wsConnecting.value = false
   })
   hot.on('vite:ws:disconnect', () => {
+    // eslint-disable-next-line no-console
+    console.log('[nuxt-devtools] Disconnected from WebSocket')
     wsConnecting.value = true
   })