+++ taze.config.ts
@@ -1,10 +1,6 @@
 import { defineConfig } from 'taze'
 
 export default defineConfig({
-  exclude: [
-    'execa', // Node compatibility,
-    'which', // Node compatibility,
-  ],
   ignorePaths: [
     'clones',
   ],